added .float() on all X, added Dropoutlayer, redused learningh rate

+++ src/utils.py
@@ -21,7 +21,7 @@ def load_train_val_loaders(
     y_path = os.path.join(data_dir, y_file)
 
     X = torch.load(x_path, map_location="cpu").float()
-    y = torch.load(y_path, map_location="cpu").long()
+    y = torch.load(y_path, map_location="cpu").float()
 
     if X.ndim != 3:
         raise ValueError(f"Expected X to have shape N x M x M, got {tuple(X.shape)}")
